Add new jd; remove old johnny_deps and generate_deps
Scripts johnny_deps and generate_deps are being deprecated, in favor of
a new tool with a different approach. jd is targeted to Godeps files
listing only first-level dependencies. The full dependencies graph is
checked for consistency as the tool downloads (using git clone) required
packages. It also does a better effort of coming up with a branch that
represents the commit listed in dependencies. (Of course it still
resorts to a detached HEAD if no better alternative is available.) Usage
is handled by subcommands, with sensible defaults for the user
convenience. For a typical build in the current directory, it's usually
enough to run "jd" with no parameters at all. See "jd help".

Closes #46.